requirements.txt
biopython
langchain
langchain-community
langchain-ollama
langchain-huggingface
langchain-text-splitters
chromadb
sentence-transformers
